[#67] refactor : SecondLevelComponent suspense 적용
